Update rest of excess length/thickness
Just dividing the excess length (or thickness) produces rounding errors, leaving unallocated pixels in cases. So, instead, update the remainder and divide again with the rest of the children. Eventually, the division will be with 1, making the last child take all the then available pixels and leaving no pixels unallocated.